requirements/dev.txt
# Developer tooling that layers on top of requirements/base.txt.
-r base.txt

# Pin toolchain versions so CI, local development, and docker images match.
mypy==1.17.1  # env_py311_2025-08-27T18-56-38Z snapshot
ruff==0.12.11
black==25.1.0
pre-commit==3.8.0
import-linter==2.5  # architecture contracts
types-PyYAML==6.0.12.20250915
pandas-stubs==2.3.2.250827
types-psutil==7.0.0.20250822
types-requests==2.32.4.20250913
scipy-stubs==1.16.2.0
pytest==8.4.1
pytest-asyncio==1.2.0
pytest-cov==7.0.0
hypothesis==6.112.0
vulture==2.11
fakeredis==2.23.2

